Removed 1 unnecessary stubbing in RemoteFileUtilsTest.java
